fix the bug of ROS , add visitor

+++ final/BL/GenerateCodeFiles/GenerateFilesUtils.cs
@@ -1,4 +1,3 @@
-
 using MongoDB.Bson;
 using System;
 using MongoDB.Bson.Serialization.Attributes;
@@ -20,48 +19,55 @@ namespace WebApiCSharp.GenerateCodeFiles
 
             var process = new Process()
             {
-            StartInfo = new ProcessStartInfo
-            {
-            FileName = "/bin/bash",
-            Arguments = $"-c \"{escapedArgs}\"",
-            RedirectStandardOutput = true,
-            UseShellExecute = false,
-            CreateNoWindow = true,
-            }
+                StartInfo = new ProcessStartInfo
+                {
+                    FileName = "/bin/bash",
+                    Arguments = $"-c \"{escapedArgs}\"",
+                    RedirectStandardOutput = true,
+                    UseShellExecute = false,
+                    CreateNoWindow = true,
+                }
             };
-            if(workingDir != null)
-            {
-              process =   new Process(){
-            StartInfo = new ProcessStartInfo
-            {
-            FileName = "/bin/bash",
-            Arguments = $"-c \"{escapedArgs}\"",
-            RedirectStandardOutput = true,
-            UseShellExecute = false,
-            CreateNoWindow = true,
-            WorkingDirectory=workingDir
-            }
-            };
-            }
             
-            process.Start();
-            if(waitForExit)
+            if (workingDir != null)
             {
-            string result = process.StandardOutput.ReadToEnd();
-            process.WaitForExit();
-
-            return result;
+                process = new Process()
+                {
+                    StartInfo = new ProcessStartInfo
+                    {
+                        FileName = "/bin/bash",
+                        Arguments = $"-c \"{escapedArgs}\"",
+                        RedirectStandardOutput = true,
+                        UseShellExecute = false,
+                        CreateNoWindow = true,
+                        WorkingDirectory = workingDir
+                    }
+                };
             }
+
+            process.Start();
+            
+            if (waitForExit)
+            {
+                string result = process.StandardOutput.ReadToEnd();
+                process.WaitForExit();
+
+                return result;
+            }
+
             return "";
         }
+
         public static void RunApplicationUntilEnd(string appFilePath, string workingDir = null, string arguments = null)
         {
             ProcessStartInfo sInfo = new ProcessStartInfo()
             {
                 FileName = appFilePath,
             };
+            
             Process process = new Process();
             process.StartInfo = sInfo;
+            
             if (workingDir != null)
             {
                 process.StartInfo.WorkingDirectory = workingDir;
@@ -71,26 +77,25 @@ namespace WebApiCSharp.GenerateCodeFiles
             {
                 process.StartInfo.Arguments = arguments;
             }
-
-
+            
             process.StartInfo.UseShellExecute = true;
             process.Start();
             process.WaitForExit();
             process.Close();
-
         }
+
         public static string AppendPath(string basePath, string pathEnd)
         {
             basePath = basePath.EndsWith("/") ? basePath.TrimEnd('/') : basePath;
             pathEnd = pathEnd.StartsWith("/") ? pathEnd.TrimStart('/') : pathEnd;
-
             string resPath = basePath + "/" + pathEnd;
+           
             return resPath;
         }
 
         public static bool IsPrimitiveType(string type, bool IncludeIsAnyValue = false)
         {
-            return PLPsData.PRIMITIVE_TYPES.Any(x => x.Equals(type)) || (IncludeIsAnyValue && type.Equals(PLPsData.ANY_VALUE_TYPE_NAME)) ;
+            return PLPsData.PRIMITIVE_TYPES.Any(x => x.Equals(type)) || (IncludeIsAnyValue && type.Equals(PLPsData.ANY_VALUE_TYPE_NAME));
         }
 
         public static string ToUpperFirstLetter(string str)
@@ -103,53 +108,38 @@ namespace WebApiCSharp.GenerateCodeFiles
         {
             return (new string(' ', numOfIndentations * indentSize));
         }
+
         public static string GetIndentationStr(int numOfIndentations, int indentSize = 4, string str = "", bool withNewLine = true, bool isPythonCode = false)
         {
             string result = GetIdent(numOfIndentations, indentSize) + str;
-            result = !isPythonCode ? result : 
-                result
-                .Replace("\n", Environment.NewLine + GetIdent(numOfIndentations, indentSize))
-                .Replace("\t", GetIdent(1, indentSize));
+            result = !isPythonCode
+                ? result
+                : result
+                    .Replace("\n", Environment.NewLine + GetIdent(numOfIndentations, indentSize))
+                    .Replace("\t", GetIdent(1, indentSize));
+            
             return result + (withNewLine ? Environment.NewLine : "");
         }
-
-
+        
         public static void WriteTextFile(string path, string content, bool setExecuteable = false)
         {
-            if (File.Exists(path))
-            {
-                File.Delete(path);
-            }
-            // Create a file to write to.
-            using (StreamWriter sw = File.CreateText(path))
-            {
-                sw.Write(content);
-            }
-            if (setExecuteable)
-            {
-                GenerateFilesUtils.RunBashCommand("chmod +x " + path);
-                //RunApplicationUntilEnd("chmod", null, "+x " + path);
-            }
+            if (File.Exists(path)) { File.Delete(path); }
+
+            using (StreamWriter sw = File.CreateText(path)) { sw.Write(content); }
+            
+            if (setExecuteable) { GenerateFilesUtils.RunBashCommand("chmod +x " + path); }
         }
 
         public static void DeleteAndCreateDirectory(string dirPath, bool createDirectory = true)
         {
-            if (Directory.Exists(dirPath))
-            {
-                Directory.Delete(dirPath, true);
-            }
-            if (createDirectory)
-            {
-                Directory.CreateDirectory(dirPath);
-            }
+            if (Directory.Exists(dirPath)) { Directory.Delete(dirPath, true); }
+
+            if (createDirectory) { Directory.CreateDirectory(dirPath); }
         }
+
         public static void DeleteDirectory2(string dirPath, bool createDirectory = true)
         {
-            if (Directory.Exists(dirPath))
-            {
-                Directory.Delete(dirPath, true);
-            }
-
+            if (Directory.Exists(dirPath)) { Directory.Delete(dirPath, true); }
         }
     }
 }
